MOTECH-2084: Moved all constants classes to the org.motechproject.task.constants package
