+++ modules/tasks/tasks/src/main/java/org/motechproject/tasks/service/TaskTriggerHandler.java
@@ -31,23 +31,23 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.motechproject.tasks.events.constants.EventDataKeys.HANDLER_ERROR_PARAM;
-import static org.motechproject.tasks.events.constants.EventDataKeys.JOB_SUBJECT;
-import static org.motechproject.tasks.events.constants.EventDataKeys.REPEAT_COUNT;
-import static org.motechproject.tasks.events.constants.EventDataKeys.REPEAT_INTERVAL_TIME;
-import static org.motechproject.tasks.events.constants.EventDataKeys.TASK_FAIL_FAILURE_DATE;
-import static org.motechproject.tasks.events.constants.EventDataKeys.TASK_FAIL_FAILURE_NUMBER;
-import static org.motechproject.tasks.events.constants.EventDataKeys.TASK_FAIL_MESSAGE;
-import static org.motechproject.tasks.events.constants.EventDataKeys.TASK_FAIL_STACK_TRACE;
-import static org.motechproject.tasks.events.constants.EventDataKeys.TASK_FAIL_TASK_ID;
-import static org.motechproject.tasks.events.constants.EventDataKeys.TASK_FAIL_TASK_NAME;
-import static org.motechproject.tasks.events.constants.EventDataKeys.TASK_FAIL_TRIGGER_DISABLED;
-import static org.motechproject.tasks.events.constants.EventDataKeys.TASK_ID;
-import static org.motechproject.tasks.events.constants.EventSubjects.SCHEDULE_REPEATING_JOB;
-import static org.motechproject.tasks.events.constants.EventSubjects.UNSCHEDULE_REPEATING_JOB;
-import static org.motechproject.tasks.events.constants.EventSubjects.createHandlerFailureSubject;
-import static org.motechproject.tasks.events.constants.EventSubjects.createHandlerSuccessSubject;
-import static org.motechproject.tasks.events.constants.TaskFailureCause.TRIGGER;
+import static org.motechproject.tasks.constants.EventDataKeys.HANDLER_ERROR_PARAM;
+import static org.motechproject.tasks.constants.EventDataKeys.JOB_SUBJECT;
+import static org.motechproject.tasks.constants.EventDataKeys.REPEAT_COUNT;
+import static org.motechproject.tasks.constants.EventDataKeys.REPEAT_INTERVAL_TIME;
+import static org.motechproject.tasks.constants.EventDataKeys.TASK_FAIL_FAILURE_DATE;
+import static org.motechproject.tasks.constants.EventDataKeys.TASK_FAIL_FAILURE_NUMBER;
+import static org.motechproject.tasks.constants.EventDataKeys.TASK_FAIL_MESSAGE;
+import static org.motechproject.tasks.constants.EventDataKeys.TASK_FAIL_STACK_TRACE;
+import static org.motechproject.tasks.constants.EventDataKeys.TASK_FAIL_TASK_ID;
+import static org.motechproject.tasks.constants.EventDataKeys.TASK_FAIL_TASK_NAME;
+import static org.motechproject.tasks.constants.EventDataKeys.TASK_FAIL_TRIGGER_DISABLED;
+import static org.motechproject.tasks.constants.EventDataKeys.TASK_ID;
+import static org.motechproject.tasks.constants.EventSubjects.SCHEDULE_REPEATING_JOB;
+import static org.motechproject.tasks.constants.EventSubjects.UNSCHEDULE_REPEATING_JOB;
+import static org.motechproject.tasks.constants.EventSubjects.createHandlerFailureSubject;
+import static org.motechproject.tasks.constants.EventSubjects.createHandlerSuccessSubject;
+import static org.motechproject.tasks.constants.TaskFailureCause.TRIGGER;
 import static org.motechproject.tasks.service.HandlerPredicates.withServiceName;
 
 /**
