+++ modules/tasks/tasks/src/test/java/org/motechproject/tasks/service/TaskTriggerHandlerTest.java
@@ -35,7 +35,7 @@ import org.motechproject.tasks.domain.TaskActivity;
 import org.motechproject.tasks.domain.TaskConfig;
 import org.motechproject.tasks.domain.TaskTriggerInformation;
 import org.motechproject.tasks.domain.TriggerEvent;
-import org.motechproject.tasks.events.constants.EventDataKeys;
+import org.motechproject.tasks.constants.EventDataKeys;
 import org.motechproject.tasks.exception.ActionNotFoundException;
 import org.motechproject.tasks.exception.TaskHandlerException;
 import org.osgi.framework.BundleContext;
@@ -91,13 +91,13 @@ import static org.motechproject.tasks.domain.ParameterType.TEXTAREA;
 import static org.motechproject.tasks.domain.ParameterType.TIME;
 import static org.motechproject.tasks.domain.ParameterType.UNICODE;
 import static org.motechproject.tasks.domain.TaskActivityType.ERROR;
-import static org.motechproject.tasks.events.constants.EventSubjects.SCHEDULE_REPEATING_JOB;
-import static org.motechproject.tasks.events.constants.EventSubjects.UNSCHEDULE_REPEATING_JOB;
-import static org.motechproject.tasks.events.constants.EventSubjects.createHandlerFailureSubject;
-import static org.motechproject.tasks.events.constants.EventSubjects.createHandlerSuccessSubject;
-import static org.motechproject.tasks.events.constants.TaskFailureCause.ACTION;
-import static org.motechproject.tasks.events.constants.TaskFailureCause.DATA_SOURCE;
-import static org.motechproject.tasks.events.constants.TaskFailureCause.TRIGGER;
+import static org.motechproject.tasks.constants.EventSubjects.SCHEDULE_REPEATING_JOB;
+import static org.motechproject.tasks.constants.EventSubjects.UNSCHEDULE_REPEATING_JOB;
+import static org.motechproject.tasks.constants.EventSubjects.createHandlerFailureSubject;
+import static org.motechproject.tasks.constants.EventSubjects.createHandlerSuccessSubject;
+import static org.motechproject.tasks.constants.TaskFailureCause.ACTION;
+import static org.motechproject.tasks.constants.TaskFailureCause.DATA_SOURCE;
+import static org.motechproject.tasks.constants.TaskFailureCause.TRIGGER;
 import static org.springframework.aop.support.AopUtils.getTargetClass;
 import static org.springframework.util.ReflectionUtils.findMethod;
